+++ Engine/SceneNodes/Model3DRenderer.cpp
@@ -124,9 +124,9 @@ BufferObject Model3DRenderer::ProcessPolygonGroup (Model* model, PolygonGroup* p
 			vertexBuffer.push_back (vertexData);
 		}
 
-		indexBuffer.push_back (3 * i);
-		indexBuffer.push_back (3 * i + 1);
-		indexBuffer.push_back (3 * i + 2);
+		indexBuffer.push_back(3 * (unsigned int)i);
+		indexBuffer.push_back(3 * (unsigned int)i + 1);
+		indexBuffer.push_back(3 * (unsigned int)i + 2);
 	}
 
 	BufferObject bufObj = BindVertexData (vertexBuffer, indexBuffer);
